simplified mongodb test configuration
